% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calculate_metrics}
\alias{calculate_metrics}
\title{Take a game object and caclulate summary statistics.}
\usage{
calculate_metrics(game_to_eval, mercy_rule)
}
\arguments{
\item{game}{a dataframe representing a game}

\item{mercy_rule_limit}{a value representing how far behind a player must get before mercy rule is applied}
}
\description{
This function takes a dataframe representing a game and calculates some summary statistics.
}

