% Generated by roxygen2 (4.0.1): do not edit by hand
\name{create_tables_from_db}
\alias{create_tables_from_db}
\title{Create a game level object from the database}
\usage{
create_tables_from_db(sql = "select * from games",
  path = file.path(Sys.getenv("HOME"), "chessDB", "chessAnalysis.db"),
  chunk_size = 1000)
}
\arguments{
\item{sql}{sql statement to select the rows of interest from the database.}

\item{path}{path to sqlite database: probably in /data of package}

\item{chunk_size}{number of rows to return at a time, essentially memory/speed trade off.}
}
\description{
This function takes a SQL query (default all): splits the result arrays and calculates
various metrics for black and white.
}

