% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_game_level_dfs}
\alias{get_game_level_dfs}
\title{Create a game level object from the database}
\usage{

  get_game_level_dfs(sql = "select games.*, p1.given_name as white_given_name, p1.surname as white_surname, \\n             p2.given_name as black_given_name, p2.surname as black_surname from games \\n             join players p1 on games.white=p1.pid join players p2 on games.black=p2.pid\\n             where processed=1",
  path = system.file(file.path("inst", "extdata", "chessAnalysis.db"), package
  = "chessdata"), FUN = NULL, limit = 1000, ...)
}
\arguments{
\item{sql}{sql statement to select the rows of interest from the database.}

\item{path}{path to sqlite database: probably in /inst/extdata/ dir of package}

\item{FUN}{function to be applied on each game dataframe. "..." as additional arguments.}

\item{limit}{maximum number of rows to return. Set to NULL for unlimited.}
}
\description{
This function takes a SQL query (default all): splits the result arrays and calculates
various metrics for black and white. Only processed rows are returned. The function applied to
each game is passed in. In most cases this function is expected to be summarize_game.
}

