% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summarize_game}
\alias{summarize_game}
\title{Take a game object and caclulate summary statistics.}
\usage{
summarize_game(game, mr_limit = 500)
}
\arguments{
\item{game}{a dataframe representing a game}

\item{mercy_rule_limit}{a value representing how far behind a player must get before mercy rule is applied}
}
\description{
This function takes a dataframe representing a game and calculates some summary statistics.
Although essentially a wrapper for calculating the 'mercy rule' and calculate_metrics.
}

